# Speed up utf8 decoding