… lib/std/unicode/UTF-8-demo.txt
… lib/std/unicode/boxes.txt