Extracted from #6390 and added `isValidCodepoint` and `utf8CountCodepoints`.
Added `unicode/benchmark.zig`.
Results for me (x86_64):

<table>
<tr>
	<td>Name
	<td>clz ReleaseFast
	<td>clz ReleaseSmall
	<td>switch ReleaseFast
	<td>switch ReleaseSmall
<tr>
	<td>short ASCII strings
	<td>332 MiB/s [3]
	<td>274 MiB/s [3]
	<td>384 MiB/s [3]
	<td>271 MiB/s [3]
<tr>
	<td>short Unicode strings
	<td>414 MiB/s [3]
	<td>523 MiB/s [3]
	<td>601 MiB/s [3]
	<td>488 MiB/s [3]
<tr>
	<td>pure ASCII strings
	<td>6729 MiB/s [80]
	<td>8990 MiB/s [80]
	<td>8501 MiB/s [80]
	<td>7485 MiB/s [80]
<tr>
	<td>pure Unicode strings
	<td>608 MiB/s [80]
	<td>648 MiB/s [80]
	<td>691 MiB/s [80]
	<td>612 MiB/s [80]
<tr>
	<td>mixed ASCII/Unicode strings
	<td>910 MiB/s [224]
	<td>1026 MiB/s [224]
	<td>973 MiB/s [224]
	<td>1094 MiB/s [224]
<tr>
	<td>'boxes.txt'
	<td>364 MiB/s [6976]
	<td>383 MiB/s [6976]
	<td>509 MiB/s [6976]
	<td>506 MiB/s [6976]
<tr>
	<td>'glasses.txt'
	<td>587 MiB/s [10014]
	<td>645 MiB/s [10014]
	<td>824 MiB/s [10014]
	<td>789 MiB/s [10014]
<tr>
	<td>'quickbrown.txt'
	<td>730 MiB/s [3285]
	<td>843 MiB/s [3285]
	<td>1145 MiB/s [3285]
	<td>1086 MiB/s [3285]
<tr>
	<td>'UTF-8-demo.txt'
	<td>363 MiB/s [7622]
	<td>385 MiB/s [7622]
	<td>470 MiB/s [7622]
	<td>470 MiB/s [7622]
</table>