@data-man 2020-11-11
<table>
<tr>
	<td>Name
	<td>using clz
	<td>using switch
	<td>using LUT
<tr>
	<td>short ASCII strings
	<td>332 MiB/s [3]
	<td>384 MiB/s [3]
	<td>358 MiB/s [3]
<tr>
	<td>short Unicode strings
	<td>414 MiB/s [3]
	<td>601 MiB/s [3]
	<td>498 MiB/s [3]
<tr>
	<td>pure ASCII strings
	<td>6729 MiB/s [80]
	<td>8501 MiB/s [80]
	<td>8512 MiB/s [80]
<tr>
	<td>pure Unicode strings
	<td>608 MiB/s [80]
	<td>691 MiB/s [80]
	<td>728 MiB/s [80]
<tr>
	<td>mixed ASCII/Unicode strings
	<td>910 MiB/s [224]
	<td>973 MiB/s [224]
	<td>690 MiB/s [224]
<tr>
	<td>'boxes.txt'
	<td>364 MiB/s [6976]
	<td>509 MiB/s [6976]
	<td>430 MiB/s [6976]
<tr>
	<td>'glasses.txt'
	<td>587 MiB/s [10014]
	<td>824 MiB/s [10014]
	<td>700 MiB/s [10014]
<tr>
	<td>'quickbrown.txt'
	<td>730 MiB/s [3285]
	<td>1145 MiB/s [3285]
	<td>960 MiB/s [3285]
<tr>
	<td>'UTF-8-demo.txt'
	<td>363 MiB/s [7622]
	<td>470 MiB/s [7622]
	<td>454 MiB/s [7622]
</table>

<details>
<summary>LUT code</summary>

```zig
const utf8_first_byte_lens = [256]u3{
    1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
    1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, // 0x1F
    1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
    1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, // 0x3F
    1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
    1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, // 0x5F
    1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
    1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, // 0x7F
    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, // 0x9F
    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
    0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, // 0xBF
    0, 0, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2,
    2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, // 0xDF
    3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, // 0xEF
    4, 4, 4, 4, 4, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, // 0xFF
};

pub fn utf8ByteSequenceLength(first_byte: u8) !u3 {
    const len = utf8_first_byte_lens[first_byte];
    if (len == 0) return error.Utf8InvalidStartByte;
    return len;
}
```
</details>