@LemonBoy 2020-11-11
Cool, you stripped away my name out of this code and lumped everything in a single commit... A true class act.

You couldn't really wait for #6390 to be reviewed?