@andrewrk 2020-11-11
> Cool, you stripped away my name out of this code and lumped everything in a single commit... A true class act.

@data-man can you please explain what happened here?