@data-man 2020-11-11
> Cool, you stripped away my name out of this code

I referenced your PR:
> Extracted from #6390

> A true class act.

I'm sorry if it hurt your feelings... :(