@data-man 2020-11-11
@LemonBoy 
> You couldn't really wait for #6390 to be reviewed?

I don't understand why you have not splitted your PR into several parts.

> // The switch is optimized much better than a "smart" approach using @clz

This comment is not a good reason. That is why I have given the real results.

@andrewrk 
> can you please explain what happened here?

I hope I have satisfied your request.