@LemonBoy 2020-11-11
> I referenced your PR:

You should've cherry-picked the commits you wanted (or apply the `.patch` file with `git am` and remove the extra commits), that's the standard way of picking-up a stalled patchset.

> I don't understand why you have not splitted your PR into several parts.

Because it was not reviewed yet?
I also have another PR touching `fmt.zig`, I was waiting for one of them to land first to rebase the other.

> This comment is not a good reason. That is why I have given the real results.

:thinking: what?
I wrote that comment after profiling the code and running a few benchmarks, what else did you expect?

> Rename utf8ValidCodepoint to isValidCodepoint 

I kept the original naming structure, the idea was to introduce some kind of namespacing a-la `crypto.zig` with `const utf8 = struct {...}` and `const utf16 = struct {...}` but was deferred to a later PR.