@data-man 2020-11-11
> You should've cherry-picked the commits you wanted 

git is still a dark forest for me. Sorry if I don't know how to do something.